split SocketBuf
